requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9
feedparser>=6.0
anthropic>=0.40
openai>=1.50
apscheduler>=3.10
python-dotenv>=1.0
trafilatura>=1.12
googlenewsdecoder>=0.1.6
fastapi>=0.111
uvicorn>=0.29
